dao_quote/requirements.txt
pandas==0.25.1
urllib3==1.24.2
gevent==1.4.0
aioredis==1.3.0
Django==2.2.6
aiosqlite==0.10.0
redis==3.3.10
aiohttp==3.6.2
requests==2.22.0
paramiko==2.7.0
pymongo==3.9.0
websocket_client==0.56.0
uWSGI==2.0.17.1
pyyaml==5.3.1
thriftpy2
tqsdk
